fetchs a api + cambios turnos profesional
