+++ data-detectives-react/src/app/types.tsx
@@ -58,10 +58,6 @@ export interface SearchPageProps {
     handleEspecialidadChange: (option: Especialidad_Profesional | null) => void;
   } 
 
-export interface ShowTurnosProfesionalProps {
-    profesional: Profesional;
-} 
-
 export interface Especialidad {
     id: number;
     nombre: string;
